requirements.txt
Django==5.0.3
django-extensions==3.2.3
djangorestframework==3.15.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
PyMySQL==1.1.0